feat : sample 폴더 제거 및 하위 폴더 clothes 폴더 하위로 경로 변경
