+++ src/main/java/com/miracle/coordifit/clothes/dto/ClothesDetailResponseDto.java
@@ -1,6 +1,7 @@
-package com.miracle.coordifit.clothes.sample.dto;
+package com.miracle.coordifit.clothes.dto;
 
 import java.time.LocalDate;
+import java.util.List;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -13,14 +14,27 @@ import lombok.Setter;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ClothesResponseSample {
+public class ClothesDetailResponseDto {
 	private String clothesId;
 	private String name;
 	private String brand;
-	private Integer price;
 	private String categoryCode;
+	private String clothesSize;
+	private Integer price;
 	private LocalDate purchaseDate;
-	private String imageUrl;
+	private String purchaseUrl;
+	private String description;
 	private Integer wearCount;
 	private LocalDate lastWornDate;
+	private List<ClothesImage> images;
+
+	@Getter
+	@Setter
+	@Builder
+	@NoArgsConstructor
+	@AllArgsConstructor
+	public static class ClothesImage {
+		private Long fileId;
+		private String url;
+	}
 }
