+++ src/main/java/com/miracle/coordifit/clothes/dto/ClothesResponseSample.java
@@ -1,7 +1,6 @@
-package com.miracle.coordifit.clothes.sample.dto;
+package com.miracle.coordifit.clothes.dto;
 
 import java.time.LocalDate;
-import java.util.List;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -14,27 +13,14 @@ import lombok.Setter;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ClothesDetailResponseDto {
+public class ClothesResponseSample {
 	private String clothesId;
 	private String name;
 	private String brand;
-	private String categoryCode;
-	private String clothesSize;
 	private Integer price;
+	private String categoryCode;
 	private LocalDate purchaseDate;
-	private String purchaseUrl;
-	private String description;
+	private String imageUrl;
 	private Integer wearCount;
 	private LocalDate lastWornDate;
-	private List<ClothesImage> images;
-
-	@Getter
-	@Setter
-	@Builder
-	@NoArgsConstructor
-	@AllArgsConstructor
-	public static class ClothesImage {
-		private Long fileId;
-		private String url;
-	}
 }
