+++ src/main/resources/mapper/clothes/ClothesMapperSample.xml
@@ -3,7 +3,7 @@
   PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.miracle.coordifit.clothes.sample.repository.ClothesRepositorySample">
+<mapper namespace="com.miracle.coordifit.clothes.repository.ClothesRepositorySample">
 
   <select id="getNextClothesSequence" resultType="int">
     SELECT NVL(MAX(TO_NUMBER(SUBSTR(clothes_id, 8))), 0) + 1
@@ -11,7 +11,7 @@
     WHERE SUBSTR(clothes_id, 2, 6) = TO_CHAR(SYSDATE, 'YYMMDD')
   </select>
 
-  <insert id="insertClothes" parameterType="com.miracle.coordifit.clothes.sample.model.ClothesSample">
+  <insert id="insertClothes" parameterType="com.miracle.coordifit.clothes.model.ClothesSample">
     INSERT INTO clothes (
       clothes_id, 
       user_id, 
@@ -49,7 +49,7 @@
     )
   </insert>
 
-  <insert id="insertClothesImage" parameterType="com.miracle.coordifit.clothes.sample.model.ClothesImageSample">
+  <insert id="insertClothesImage" parameterType="com.miracle.coordifit.clothes.model.ClothesImageSample">
     INSERT INTO clothes_images (
       clothes_id, file_id, created_at, created_by
     ) VALUES (
@@ -57,7 +57,7 @@
     )
   </insert>
 
-  <select id="selectUserClothes" parameterType="string" resultType="com.miracle.coordifit.clothes.sample.dto.ClothesResponseSample">
+  <select id="selectUserClothes" parameterType="string" resultType="com.miracle.coordifit.clothes.dto.ClothesResponseSample">
     SELECT 
       c.clothes_id AS clothesId,
       c.name AS name,
@@ -82,7 +82,7 @@
     ORDER BY c.created_at DESC
   </select>
 
-  <select id="selectClothesById" resultType="com.miracle.coordifit.clothes.sample.dto.ClothesDetailResponseDto">
+  <select id="selectClothesById" resultType="com.miracle.coordifit.clothes.dto.ClothesDetailResponseDto">
     SELECT 
       clothes_id as clothesId,
       name,
@@ -101,7 +101,7 @@
     AND is_active = 'Y'
   </select>
 
-  <select id="selectClothesImage" resultType="com.miracle.coordifit.clothes.sample.dto.ClothesDetailResponseDto$ClothesImage">
+  <select id="selectClothesImage" resultType="com.miracle.coordifit.clothes.dto.ClothesDetailResponseDto$ClothesImage">
     SELECT 
       ci.file_id AS fileId,
       f.s3_url AS url
@@ -117,7 +117,7 @@
     AND file_id = #{fileId}
   </delete>
 
-  <update id="updateClothes" parameterType="com.miracle.coordifit.clothes.sample.model.ClothesSample">
+  <update id="updateClothes" parameterType="com.miracle.coordifit.clothes.model.ClothesSample">
     UPDATE clothes
     SET 
       name = #{name},
@@ -133,7 +133,7 @@
     WHERE clothes_id = #{clothesId}
   </update>
 
-  <update id="deleteClothes" parameterType="com.miracle.coordifit.clothes.sample.model.ClothesSample">
+  <update id="deleteClothes" parameterType="com.miracle.coordifit.clothes.model.ClothesSample">
     UPDATE clothes
     SET 
       is_active = 'N',
